backend/python-ai/requirements.txt
# Core dependencies for summary evaluation
sentence-transformers>=2.2.0
flask>=2.3.0
flask-cors>=4.0.0
requests>=2.31.0
python-dotenv>=1.0.0
numpy>=1.24.0
